LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-referer-expected.txt


FAIL XMLHttpRequest: open() - value of Referer header assert_equals: expected "Referer: http://localhost:8800/XMLHttpRequest/open-referer.htm\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"/Volumes/Data/EWS/WebKit/LayoutTests/imported/w3c/web-platform-tests/tools/wptserve/wptserve/handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"/Volumes/Data/EWS/WebKit/LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
